Implemented - Player Face Enemy Rotation
This commit implements functionality for the player to rotate to face the selected enemy on click of enemy icon.

Functionality has also been added to have the camera correctly adjust to still be behind the player when a new enemy is selected.

Next, continue to work on this functionality and maybe focus on having the grid be disabled when in the enemy select mode.