Implemented - OutOfRange Enemies Can't be Selected
This commit implements functionality to prevent enemies that are out of range of the player's weapon from being selected. This helps the player to recognise which enemies can be attacked.

The commit also fixes an issue where the player's actions wouldn't be re-enabled if the player was to exit out of the mode when in transition.

Next, add a symbol above the enemyIcon so the player can better understand which enemy relates to which icon.