Added/Implemented - Melee Anims + Refactored DMG Calcs
This commit first adds two different melee animations and adds the basic set-up for the melee ability methods. It also refactors the Fire, SteadyAim, and QuickDraw methods so dmg calc, application, and effect plays are done in ONE predefined method instead of being repeated in each ability method.

Overall this commit aimed to improve the readability of the PlayerActions Script as well as the performance.

Next, continue work implementing the Melee ability and also fix the issue with Follow Up Shot being allowed to quickly after QuickDraw has been used.