Implemented - SelectedEnemy Icon
This commit implements functionality to show the new SelectedEnemy Icon above the Enemy Icon whenever the assignedEnemy is selected. AssignedEnemy is being used so if the player uses hover click to select the enemy rather than the enemy icon they can still easily see which icon is associated to it.

Next, implement a way to have the enemy attacked be the selectedEnemy rather than closest. Closest should now never be used as when clicking onto an action it should also enter this mode.