Implemented - QuickDraw & FollowUpShot Merge
This commit merges QuickDrawn and FollowUpShot into one space on the actionBar, when QuickDraw has been used it is replaced with FollowUpShot which can be used for a futher lower accurarcy attack.

FollowUpShot goes away upon use and is re-replaced with QuickDraw at the end of the turn if not already used.

Next, add the Melee ability icon to the actionBar in place of the old FollowUpShot.