Implemented - Melee ActionBar Icon
This commit adds the new Melee ActionBar Icon in place of where FollowUpShot used to be, it also implements all the changes needed to have it working as a functional part of the UI as well as the neuonces for it being a Melee ability rather than a standard shot one.

Next, merge what we have here into main as the functionality for the Melee ability and refactored PlayerActions is complete.