Implemented - ActionBar Abilities Enter EnemySelect View
This commit implements functionality so that when clicking on an ability when in normal view the cam is switched to enemySelect view to give the player the ability to select an enemy rather than a random one being chosen for them.

They must then click an action again to attack their selected enemy.

Next, get this enemySelect view entering working for both Melee and QuickDraw & FollowUpShot.