Fixed - Varriance in EnemySelect Entering Time
This commit fixes the issues that were causing the EnemySelect view to be entered at differeing times depending on a number of factors (pos, rot, frames, etc). It makes the view/mode feel nicer and more consistent to use.

Next, have the enemySelect icons move depending on if a background info is shown or not.