Implemented - HealthStation Grid Nodes
This commit implements the needed functionality for HealthStations to appear as green nodes in gridVisuals.

Some placeholder text also appears in the UI to prompt the player to click E to use.

Next, allow the player to use the healthStation when they click E. Need to find out if they can use the healthStation again or if it's one time use, and what amount of health is restored.