Implemented - TurretAI Scripts
This commit finally gets to work on making the Turret enemy type functional, currently the camera will just pan to the turret when it's their turn but other than implementing the attack stuff everythings all their for it.

It's offically a functioning enemy type woo!

Next, finish implementing it's attack stuff and make sure the player can kill it.