Implemented - Full Turret
This commit re-implements alot of what I did yesterday as I decided it's better to have the turret on the grid as opposed to a seperate entity.

Everything works now just needs polishing up and the turret shoot sfx adding.

Next, polish up the turret stuff and then move on to the rest of level 3/boss fight.